==> 2.P337_BAL_BE_gene_analysis.Rmd <==
---
title: "P337: Compare gene expression in BAL and BE samples"
author: "Kim Dill-McFarland, kadm@uw.edu"
output:
  html_document:
    toc: yes
    toc_depth: 3
    toc_float:
      collapsed: no
date: "version `r format(Sys.time(), '%B %d, %Y')`"
editor_options: 
  chunk_output_type: console
---

# Background

The purpose of this workflow is to define the epithelial and cellular expression patterns that relate to eosinophil recruitment into the airway after allergen challenge. 

These data were previously cleaned and modules were built from genes significant for allergen challenges. Briefly, low quality libraries were removed (median CV > 0.65, alignment < 80%, sequences < 1E6). Genes were filtered to protein-coding expressed at > 1 CPM in at least 10% of BAL and BE samples separately. TMM and voom log2 normalization was then completed. For details, see:

* BAL: <https://github.com/altman-lab/P337_MINA_BAL_public> 
* BE: <https://github.com/altman-lab/P337_MINA_BE_public>

# Setup

Load packages

```{r message=FALSE, warning=FALSE}
# Data manipulation and figures
library(tidyverse)
# RNAseq data
library(limma)
library(kimma)
#plots
library(BIGpicture)
library(patchwork)
# Print pretty table to knit file
library(knitr)
library(kableExtra)
options(knitr.kable.NA = '')

`%notin%` <- Negate(`%in%`)
```

Set seed

```{r}
set.seed(4389)
```

# Data
## Load data

```{r}
# Genes
attach("data_clean/P337_BAL_data.RData")
attach("data_clean/P337_BE_data.RData")
```

Recode visit

```{r}
dat.BAL$targets <- dat.BAL$targets %>% 
  mutate(visit = case_when(visit == "V4" ~ "Pre",
                           visit == "V5" ~ "Post"),
         visit = factor(visit, levels = c("Pre","Post")))

dat.BE$targets <- dat.BE$targets %>% 
  mutate(visit = case_when(visit == "V4" ~ "Pre",
                           visit == "V5" ~ "Post"),
         visit = factor(visit, levels = c("Pre","Post")))
```

# Linear modeling

```{r echo=FALSE}
load("results/gene_model_fitting.RData")
```

## Linear model: visit

#### BAL

```{r eval=FALSE}
# Define model
model.visitA <- model.matrix(~ visit, data=dat.BAL$targets)
  colnames(model.visitA) <- c("(Intercept)", "visit")
  
#block by donor
consensus.corrA <- duplicateCorrelation(
                    dat.BAL$E,
                    model.visitA,
  block=dat.BAL$targets$donorID)$consensus.correlation
  
consensus.corrA
  
# Fit model to transformed count data. Calculate eBayes
efitA <- eBayes(
            lmFit(dat.BAL$E, model.visitA,
                  block=dat.BAL$targets$donorID,
                  correlation=consensus.corrA))

#Extract results
lm_visit_BAL <- extract_lmFit(model.visitA, efitA)
```

#### BE

```{r eval=FALSE}
# Define model
model.visitE <- model.matrix(~ visit, data=dat.BE$targets)
  colnames(model.visitE) <- c("(Intercept)", "visit")
  
#block by donor
consensus.corrE <- duplicateCorrelation(
                    dat.BE$E,
                    model.visitE,
  block=dat.BE$targets$donorID)$consensus.correlation
  
consensus.corrE
  
# Fit model to transformed count data. Calculate eBayes
efitE <- eBayes(
            lmFit(dat.BE$E, model.visitE,
                  block=dat.BE$targets$donorID,
                  correlation=consensus.corrE))

#Extract results
lm_visit_BE <- extract_lmFit(model.visitE, efitE)
```

### Summarize visit models

```{r echo=FALSE}
summarise_kmFit(lm_visit_BAL$lm) %>% 
  kable(caption = "Post vs Pre visit in BAL") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_visit_BE$lm) %>% 
  kable(caption = "Post vs Pre visit in BE") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```

## Linear models within visit significant genes

Test association with EOS and PMN percentages in only visit-significant genes in Post-visit samples.

### Select visit significant genes and "Post" samples

```{r}
#Maximum fdr for visit genes to be included in modules
visit.fdr.cutoff <- 0.3
```

#### BAL

```{r}
#Subset data to visit signif genes
##List genes
visit.signif.BAL <- lm_visit_BAL$lm %>% 
  filter(FDR <= visit.fdr.cutoff & variable == "visit") %>% 
  pull(gene)
  
##Subset voom data
dat.BAL.visit <- dat.BAL

dat.BAL.visit$targets <- dat.BAL.visit$targets %>% 
  filter(visit == "Post")

dat.BAL.visit$E <- as.data.frame(dat.BAL.visit$E) %>% 
  rownames_to_column() %>% 
  filter(rowname %in% visit.signif.BAL) %>% 
  column_to_rownames() %>% 
  select(all_of(dat.BAL.visit$targets$libID))

dat.BAL.visit$genes <- as.data.frame(dat.BAL.visit$genes) %>% 
  filter(geneName %in% visit.signif.BAL)

rownames(dat.BAL.visit$weights) <- rownames(dat.BAL$E)
colnames(dat.BAL.visit$weights) <- colnames(dat.BAL$E)
dat.BAL.visit$weights <- as.data.frame(dat.BAL.visit$weights) %>% 
  rownames_to_column() %>% 
  filter(rowname %in% visit.signif.BAL) %>% 
  column_to_rownames() %>% 
  select(all_of(dat.BAL.visit$targets$libID))
```

#### BE

```{r}
#Subset data to visit signif genes
##List genes
visit.signif.BE <- lm_visit_BE$lm %>% 
  filter(FDR <= visit.fdr.cutoff & variable == "visit") %>% 
  pull(gene)
  
##Subset voom data
dat.BE.visit <- dat.BE

dat.BE.visit$targets <- dat.BE.visit$targets %>% 
  filter(visit == "Post")

dat.BE.visit$E <- as.data.frame(dat.BE.visit$E) %>% 
  rownames_to_column() %>% 
  filter(rowname %in% visit.signif.BE) %>% 
  column_to_rownames() %>% 
  select(all_of(dat.BE.visit$targets$libID))

dat.BE.visit$genes <- as.data.frame(dat.BE.visit$genes) %>% 
  filter(geneName %in% visit.signif.BE)

rownames(dat.BE.visit$weights) <- rownames(dat.BE$E)
colnames(dat.BE.visit$weights) <- colnames(dat.BE$E)
dat.BE.visit$weights <- as.data.frame(dat.BE.visit$weights) %>% 
  rownames_to_column() %>% 
  filter(rowname %in% visit.signif.BE) %>% 
  column_to_rownames() %>% 
  select(all_of(dat.BE.visit$targets$libID))
```

#### Check sizes

```{r}
dim(dat.BAL.visit)
dim(dat.BE.visit)
```

### Linear model: EOS
#### BAL

```{r eval=FALSE}
# Define model
model.eosA <- model.matrix(~ EOS.pct, data=dat.BAL.visit$targets)

#No blocking as there is one post sample per donor
  
# Fit model to transformed count data. Calculate eBayes
efitA2 <- eBayes(
            lmFit(dat.BAL.visit$E, model.eosA))

#Extract results
lm_eos_BAL <- extract_lmFit(model.eosA, efitA2)
```

#### BE

```{r eval=FALSE}
# Define model
model.eosE <- model.matrix(~ EOS.pct, data=dat.BE.visit$targets)

#No blocking as there is one post sample per donor

# Fit model to transformed count data. Calculate eBayes
efitE2 <- eBayes(
            lmFit(dat.BE.visit$E, model.eosE))

#Extract results
lm_eos_BE <- extract_lmFit(model.eosE, efitE2)
```

#### Summarize EOS models

```{r echo=FALSE}
summarise_kmFit(lm_eos_BAL$lm) %>% 
  kable(caption = "Percent EOS in BAL, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_eos_BE$lm) %>% 
  kable(caption = "Percent EOS in BE, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```

### Linear model: PMN
#### BAL

```{r eval=FALSE}
# Define model
model.neutA <- model.matrix(~ PMN.pct, data=dat.BAL.visit$targets)

#No blocking as there is one post sample per donor
  
# Fit model to transformed count data. Calculate eBayes
efitA3 <- eBayes(
            lmFit(dat.BAL.visit$E, model.neutA))

#Extract results
lm_neut_BAL <- extract_lmFit(model.neutA, efitA3)
```

#### BE

```{r eval=FALSE}
# Define model
model.neutE <- model.matrix(~ PMN.pct, data=dat.BE.visit$targets)

#No blocking as there is one post sample per donor

# Fit model to transformed count data. Calculate eBayes
efitE3 <- eBayes(
            lmFit(dat.BE.visit$E, model.neutE))

#Extract results
lm_neut_BE <- extract_lmFit(model.neutE, efitE3)
```

#### Summarize PMN models

```{r echo=FALSE, warning=FALSE, error=TRUE}
summarise_kmFit(lm_neut_BAL$lm) %>% 
  kable(caption = "Percent PMN in BAL, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_neut_BE$lm) %>% 
  kable(caption = "Percent PMN in BE, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```

### Linear model: Delta PMN

Because neutrophils vary at baseline compare the post-challenge model to one with the change in neutrophils

```{r}
#Calculate delta neut
neut.delta.bal <- dat.BAL$targets %>% 
  select(donorID, visit, PMN.pct) %>% 
  pivot_wider(names_from = visit, values_from = PMN.pct) %>% 
  mutate(PMN.delta = Post-Pre)

dat.BAL.visit$targets <- dat.BAL.visit$targets %>% 
  left_join(neut.delta.bal)
####
neut.delta.be <- dat.BE$targets %>% 
  select(donorID, visit, PMN.pct) %>% 
  pivot_wider(names_from = visit, values_from = PMN.pct) %>% 
  mutate(PMN.delta = Post-Pre)

dat.BE.visit$targets <- dat.BE.visit$targets %>% 
  left_join(neut.delta.be)
```

#### BAL

```{r eval=FALSE}
# Define model
model.neutA2 <- model.matrix(~ PMN.delta, data=dat.BAL.visit$targets)

#No blocking as there is one post sample per donor
  
# Fit model to transformed count data. Calculate eBayes
efitA4 <- eBayes(
            lmFit(dat.BAL.visit$E, model.neutA2))

#Extract results
lm_neut_delta_BAL <- extract_lmFit(model.neutA2, efitA4)
```

#### BE

```{r eval=FALSE}
#remove missing sample
dat.BE.visit.sub <- dat.BE.visit
dat.BE.visit.sub$targets <- dat.BE.visit.sub$targets %>% 
  drop_na(PMN.delta)
dat.BE.visit.sub$E <- as.data.frame(dat.BE.visit.sub$E) %>% 
  select(all_of(dat.BE.visit.sub$targets$libID))
dat.BE.visit.sub$weights <- dat.BE.visit.sub$weights %>% 
  select(all_of(dat.BE.visit.sub$targets$libID))

# Define model
model.neutE2 <- model.matrix(~ PMN.delta, data=dat.BE.visit.sub$targets)

#No blocking as there is one post sample per donor

# Fit model to transformed count data. Calculate eBayes
efitE4 <- eBayes(
            lmFit(dat.BE.visit.sub$E, model.neutE2))

#Extract results
lm_neut_delta_BE <- extract_lmFit(model.neutE2, efitE4)
```

#### Summarize delta PMN models

```{r echo=FALSE, warning=FALSE, error=TRUE}
summarise_kmFit(lm_neut_delta_BAL$lm) %>% 
  kable(caption = "Percent PMN in BAL, Change from Pre to Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_neut_delta_BE$lm) %>% 
  kable(caption = "Percent PMN in BE, Change from Pre to Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```

Compare to non-delta model. No genes significant for either PMN variable in BE

```{r}
plot_venn_genes(list("PMN BAL"=lm_neut_BAL$lm,
                     "PMN delta BAL"=lm_neut_delta_BAL$lm),
                fdr_cutoff = 0.1, return_genes = TRUE)

plot_venn_genes(list("PMN BE"=lm_neut_BE$lm,
                     "PMN delta BE"=lm_neut_delta_BE$lm),
                fdr_cutoff = 0.1)
```

### Linear model: FeNO

Test association with FeNO in only visit-significant genes in Post-visit samples.

#### BAL

Run models in kimma to get fit metrics and determine if FeNO log transformation is needed. 

```{r}
lm_feno_BAL_k1 <- kmFit(dat.BAL.visit, patientID = "donorID",
                        model = "~FeNO.PreBro_V5", run_lm = TRUE,
                        metrics = TRUE)

dat.BAL.visit$targets <- dat.BAL.visit$targets %>% 
  mutate(log_FeNO.PreBro_V5 = log2(FeNO.PreBro_V5))

lm_feno_BAL_k2 <- kmFit(dat.BAL.visit, patientID = "donorID",
                        model = "~log_FeNO.PreBro_V5", run_lm = TRUE,
                        metrics = TRUE)

plot_fit2(lm_feno_BAL_k1, x = "lm",
          lm_feno_BAL_k2, y = "lm")

lm_feno_BAL_k1$lm.fit %>% 
  mutate(model="FeNO") %>% 
  bind_rows(lm_feno_BAL_k2$lm.fit %>% mutate(model="log_FeNO")) %>% 
  select(model, gene, AIC) %>% 
  pivot_wider(names_from = model, values_from = AIC) %>% 
  mutate(deltaAIC = log_FeNO-FeNO) %>% 
  mutate(`best_fit_AIC>2` = case_when(deltaAIC < -2 ~ "log_FeNO", 
                                      deltaAIC > 2 ~ "FeNO",
                                      TRUE ~ "none")) %>% 
  count(`best_fit_AIC>2`)
```

A non-log model better fits more genes, though differences are very small.

```{r eval=FALSE}
# Define model
model.fenoA <- model.matrix(~ FeNO.PreBro_V5, data=dat.BAL.visit$targets)
  colnames(model.fenoA) <- c("(Intercept)", "FeNO")

#No blocking as there is one post sample per donor
  
# Fit model to transformed count data. Calculate eBayes
efitA3 <- eBayes(
            lmFit(dat.BAL.visit$E, model.fenoA))

#Extract results
lm_feno_BAL <- extract_lmFit(model.fenoA, efitA3)
```

#### BE

```{r eval=FALSE}
# Define model
model.fenoE <- model.matrix(~ FeNO.PreBro_V5, data=dat.BE.visit$targets)
  colnames(model.fenoE) <- c("(Intercept)", "FeNO")

#No blocking as there is one post sample per donor

# Fit model to transformed count data. Calculate eBayes
efitE3 <- eBayes(
            lmFit(dat.BE.visit$E, model.fenoE))

#Extract results
lm_feno_BE <- extract_lmFit(model.fenoE, efitE3)
```

#### Summarize FeNO models

```{r echo=FALSE}
summarise_kmFit(lm_feno_BAL$lm) %>% 
  kable(caption = "FeNO in BAL, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_feno_BE$lm) %>% 
  kable(caption = "FeNO in BE, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```


### Linear model: FEV1

Test association with FEV1 in only visit-significant genes in Post-visit samples.

#### BAL

```{r eval=FALSE}
# Define model
model.fevA <- model.matrix(~ FEV1.pctPP.preAlbuterol_V5, data=dat.BAL.visit$targets)
  colnames(model.fevA) <- c("(Intercept)", "FEV1")

#No blocking as there is one post sample per donor
  
# Fit model to transformed count data. Calculate eBayes
efitV <- eBayes(
            lmFit(dat.BAL.visit$E, model.fevA))

#Extract results
lm_fev_BAL <- extract_lmFit(model.fevA, efitV)
```

#### BE

```{r eval=FALSE}
# Define model
model.fevE <- model.matrix(~ FEV1.pctPP.preAlbuterol_V5, data=dat.BE.visit$targets)
  colnames(model.fevE) <- c("(Intercept)", "FEV1")

#No blocking as there is one post sample per donor

# Fit model to transformed count data. Calculate eBayes
efitV2 <- eBayes(
            lmFit(dat.BE.visit$E, model.fevE))

#Extract results
lm_fev_BE <- extract_lmFit(model.fevE, efitV2)
```

#### Summarize FEV1 models

```{r echo=FALSE, error=TRUE}
summarise_kmFit(lm_fev_BAL$lm) %>% 
  kable(caption = "FEV1 in BAL, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_fev_BE$lm) %>% 
  kable(caption = "FEV1 in BE, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```

### Linear model: FEV1 group

Test association with FEV1 in only visit-significant genes in Post-visit samples.

#### BAL

```{r}
dat.BAL.visit$targets <- dat.BAL.visit$targets %>% 
  mutate(FEV1.group = ifelse(FEV1.pctPP.preAlbuterol_V5 <=87, "low","high"),
         FEV1.group = factor(FEV1.group, levels=c("low","high")))

dat.BE.visit$targets <- dat.BE.visit$targets %>% 
  mutate(FEV1.group = ifelse(FEV1.pctPP.preAlbuterol_V5 <=87, "low","high"),
         FEV1.group = factor(FEV1.group, levels=c("low","high")))
```

```{r eval=FALSE}
# Define model
model.fevA2 <- model.matrix(~ FEV1.group, data=dat.BAL.visit$targets)

#No blocking as there is one post sample per donor
  
# Fit model to transformed count data. Calculate eBayes
efitV2 <- eBayes(
            lmFit(dat.BAL.visit$E, model.fevA2))

#Extract results
lm_fev_group_BAL <- extract_lmFit(model.fevA2, efitV2)
```

#### BE

```{r eval=FALSE}
# Define model
model.fevE2 <- model.matrix(~ FEV1.group, data=dat.BE.visit$targets)

#No blocking as there is one post sample per donor

# Fit model to transformed count data. Calculate eBayes
efitV3 <- eBayes(
            lmFit(dat.BE.visit$E, model.fevE2))

#Extract results
lm_fev_group_BE <- extract_lmFit(model.fevE, efitV3)
```

#### Summarize FEV1 models

```{r echo=FALSE, error=TRUE}
summarise_kmFit(lm_fev_group_BAL$lm) %>% 
  kable(caption = "FEV1 in BAL, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)

summarise_kmFit(lm_fev_group_BE$lm) %>% 
  kable(caption = "FEV1 in BE, Post samples") %>% 
  kable_styling(bootstrap_options = "striped", full_width = FALSE)
```

## Compare significant genes

```{r fig.width=8.5, echo=FALSE}
plot_upset_genes(model_result = list("BAL EOS"=lm_eos_BAL$lm,
                                    "BAL PMN"=lm_neut_BAL$lm,
                                    "BAL FeNO"=lm_feno_BAL$lm),
                variables = c("EOS.pct","FeNO","PMN.pct"),
                fdr.cutoff = c(0.1,0.3))$upset %>% 
  wrap_plots() + plot_annotation(title="BAL")

plot_upset_genes(model_result = list("BE EOS"=lm_eos_BE$lm,
                                    "BE PMN"=lm_neut_BE$lm,
                                    "BE FeNO"=lm_feno_BE$lm),
                variables = c("EOS.pct","FeNO","PMN.pct"),
                fdr.cutoff = c(0.1,0.3))$upset %>% 
  wrap_plots() + plot_annotation(title="BE")
```

## Save model results

```{r eval=FALSE}
gene.key <- dat.BAL$genes %>% 
  bind_rows(dat.BE$genes) %>% 
  select(geneName, hgnc_symbol) %>% 
  distinct()

save(lm_visit_BAL, lm_visit_BE, 
     lm_eos_BAL, lm_eos_BE, 
     lm_neut_BAL, lm_neut_BE, 
     lm_neut_delta_BAL, lm_neut_delta_BE,
     lm_feno_BAL, lm_feno_BE,
     lm_fev_BAL, lm_fev_BE,
     lm_fev_group_BAL, lm_fev_group_BE,
     file = "results/gene_model_fitting.RData")

bind_rows(lm_visit_BAL$lm, lm_eos_BAL$lm, 
          lm_neut_BAL$lm, lm_feno_BAL$lm) %>% 
  mutate(sample = "BAL") %>%  
  bind_rows(lm_visit_BE$lm, lm_eos_BE$lm, 
            lm_neut_BE$lm, lm_feno_BE$lm %>% 
              mutate(sample = "BE")) %>%
  filter(variable != "(Intercept)") %>% 
  left_join(gene.key, by = c("gene"="geneName")) %>% 
  select(sample, gene, hgnc_symbol, variable, estimate, pval, FDR) %>% 
  write_csv("results/gene_lm_results.csv")
```

# R session

```{r}
sessionInfo()
```

***
